Join public room to own module
