+++ app/page.tsx
@@ -2,7 +2,7 @@ import prisma from "@/lib/prisma";
 import roomService from "@/lib/roomService";
 
 import CreateRoom from "./room/createRoom";
-import JoinRoomButton from "./room/joinRoomButton";
+import JoinPublicRoom from "./room/joinPublicRoom";
 import JoinPrivateRoom from "./room/joinPrivateRoom";
 
 export default async function Home() {
@@ -16,18 +16,13 @@ export default async function Home() {
       <section className="flex flex-wrap gap-2">
         {publicRooms.map((room) => {
           return (
-            <div
-              key={room.id}
-              className="border w-64 p-5 rounded border-gray-300"
-            >
-              <h2 className="text-xl font-medium">{room.name}</h2>
-              <p>
-                Particpents count:{" "}
-                {rooms.find((liveRoom) => liveRoom.name == room.id.toString())
-                  ?.numParticipants || 0}
-              </p>
-              <JoinRoomButton room={room} />
-            </div>
+            <JoinPublicRoom
+              room={room}
+              participantsCount={
+                rooms.find((liveRoom) => liveRoom.name == room.id.toString())
+                  ?.numParticipants || 0
+              }
+            />
           );
         })}
       </section>
